Store and display fractional waves
TourneyRow.wave becomes a FloatField (migration 0044) so the v29
tiebreaker survives import: with an integer column the fraction was
floored on save, and reposition() then re-derived positions from the
floored ties, disagreeing with the positions computed at import time.
The live archive floored waves the same way while deduplicating and
verifying snapshots; it now carries floats end to end.

Whole waves keep rendering as integers: a shared format_wave helper is
applied wherever a wave column goes through a pandas Styler (which
would otherwise print 1234.000000) and in the overview and live-results
text, showing at most four decimals and only when present.

On SQLite the migration rebuilds tourneyrow and its history table, so
run it in a quiet window with the importer and recalc worker paused.

+++ src/thetower/backend/tourney_results/data.py
@@ -50,7 +50,7 @@ def date_to_patch(date: datetime.datetime) -> Optional[Patch]:
 
 
 @ttl_cache(maxsize=512, ttl=60)
-def wave_to_role_in_patch(roles: list[Role], role_bot_top: list[tuple[int, int]], wave: int) -> Optional[Role]:
+def wave_to_role_in_patch(roles: list[Role], role_bot_top: list[tuple[int, int]], wave: float) -> Optional[Role]:
     for role, (wave_bottom, wave_top) in zip(roles, role_bot_top):
         if wave >= wave_bottom and wave < wave_top:
             return role
@@ -80,7 +80,7 @@ def patch_to_roles(league):
     return patch_to_roles
 
 
-def wave_to_role(wave: int, patch: Optional[Patch], league: str) -> Optional[Role]:
+def wave_to_role(wave: float, patch: Optional[Patch], league: str) -> Optional[Role]:
     if not patch:
         return None
